// service/api/user/model/user-member-singup-request-data-model.ts
export default interface UserMemberSingupRequestDataModel {
  adultOk: boolean;
  authId: string;
  collectionOfMarketingInfoOk: boolean;
  collectionOfPersonalInfoOk: boolean;
  email: string;
  nickname: string;
  password: string;
  phoneNum: string;
  serviceOk: boolean;
  snsType: string;
  username: string;
}